# Conflicts:
#	Coupon System Web-App/.classpath
#	Coupon System Web-App/src/com/ronhagai/couponfaphase3/rest/controller/CouponRestController.java
#	Coupon System Web-App/src/com/ronhagai/couponfaphase3/rest/controller/LoginRestController.java

+++ Coupon System Web-App/src/com/ronhagai/couponfaphase3/core/service/CompanyService.java
@@ -9,6 +9,7 @@ import com.ronhagai.couponfaphase3.core.dao.CompanyDAO;
 import com.ronhagai.couponfaphase3.core.dao.CouponDAO;
 import com.ronhagai.couponfaphase3.core.dao.ICompanyDAO;
 import com.ronhagai.couponfaphase3.core.dao.ICouponDAO;
+import com.ronhagai.couponfaphase3.core.enums.ClientType;
 import com.ronhagai.couponfaphase3.core.exception.CouponSystemException;
 import com.ronhagai.couponfaphase3.core.exception.ExceptionsEnum;
 import com.ronhagai.couponfaphase3.core.util.ConnectionPool;
@@ -51,13 +52,19 @@ public class CompanyService implements Serializable, IBeanValidatorConstants{
 	 *  If company name or ID already exist in the DB
 	 *  If company creation fails
 	 */
-	public long createCompany(CompanyBean company) throws CouponSystemException {
+	public long createCompany(CompanyBean company/*, long userId, ClientType userType*/) throws CouponSystemException {
+		//can be used if company registration should be restricted 
+		/*if (!userType.equals(ClientType.ADMIN)) {
+			throw new CouponSystemException(ExceptionsEnum.SECURITY_BREACH,"User " + userType + " - " + userId + " attempts to create company " + company);
+		}*/
 		checkCompany(company);
 		//CLD BE HANDLED BY DAO LAYER BY MAKING IT UNIQUE
 		if(companyDAO.companyNameAlreadyExists(company.getCompName())) {
 			throw new CouponSystemException(ExceptionsEnum.NAME_EXISTS,"Company Name already exists");
 		}	
-		return companyDAO.createCompany(company);		
+		long companyId = companyDAO.createCompany(company);
+//		System.out.println("LOG : User " + userType + " - " + userId + " created company " + company);
+		return companyId;
 	}
 
 
@@ -68,18 +75,29 @@ public class CompanyService implements Serializable, IBeanValidatorConstants{
 	 *  If given company's details don't match the details of the company with the same ID in the DB
 	 *  If updating of the company's details fails
 	 */
-	public void updateCompany(CompanyBean company) throws CouponSystemException {
+	public void updateCompany(CompanyBean company, long userId, ClientType userType) throws CouponSystemException {
+		if ((company.getId() != userId || !userType.equals(ClientType.COMPANY)) && !userType.equals(ClientType.ADMIN)) {
+			throw new CouponSystemException(ExceptionsEnum.SECURITY_BREACH,"User " + userType + " - " + userId + " attempts to update company " + company);
+		}
 		checkCompany(company);
-		
 		CompanyBean tmpCompany = getCompany(company.getId());
 		tmpCompany.setEmail(company.getEmail());
 		tmpCompany.setPassword(company.getPassword());
 		
 		companyDAO.updateCompany(tmpCompany);
+		System.out.println("LOG : User " + userType + " - " + userId + " updated company " + company);		
 	}
 
-	public void updateCompanyPassword(long companyId, String oldPassword, String newPassword) throws CouponSystemException {
-		companyDAO.updateCompanyPassword(companyId, oldPassword, newPassword);
+	public void updateCompanyPassword(long companyId, String oldPassword, String newPassword, long userId, ClientType userType) throws CouponSystemException {
+		if ((companyId != userId || !userType.equals(ClientType.COMPANY)) && !userType.equals(ClientType.ADMIN)) {
+			throw new CouponSystemException(ExceptionsEnum.SECURITY_BREACH,"User " + userType + " - " + userId + " attempts to change company's password " + companyId);
+		}
+		if(userType.equals(ClientType.CUSTOMER)) {			
+			String customerName = companyDAO.getCompany(companyId).getCompName();
+			companyDAO.companyLogin(customerName, oldPassword);
+		}	
+		companyDAO.updateCompanyPassword(companyId, newPassword);
+		System.out.println("LOG : User " + userType + " - " + userId + " changed company's password " + companyId);		
 	}
 	/**
 	 * Deletes a company in the database
@@ -89,7 +107,12 @@ public class CompanyService implements Serializable, IBeanValidatorConstants{
 	 *  If given company's details don't match the details of the company with the same ID in the DB
 	 *  If company deletion fails
 	 */
-	public void removeCompany(long companyId) throws CouponSystemException {
+	public void removeCompany(long companyId, long userId, ClientType userType) throws CouponSystemException {
+		//can be modified if company removing should be restricted
+		if ((companyId != userId || !userType.equals(ClientType.COMPANY)) && !userType.equals(ClientType.ADMIN)) {
+//		if (!userType.equals(ClientType.ADMIN)) {
+			throw new CouponSystemException(ExceptionsEnum.SECURITY_BREACH,"User " + userType + " - " + userId + " attempts to remove company " + companyId);
+		}
 		connectionPool.startTransaction();
 		try {	
 			couponDAO.removeCompanyCouponsFromCustomers(companyId);
@@ -101,7 +124,7 @@ public class CompanyService implements Serializable, IBeanValidatorConstants{
 		} finally {
 		}
 		connectionPool.endTransaction();
-	
+		System.out.println("LOG : User " + userType + " - " + userId + " removed company " + companyId);			
 	}
